2024.3.6;error occurred while applying a result map. setting property 'setUserid' of 'BookMark(id=5, name=宇宙起源, url=https://www.youtube.com, userid=0)'.  Cause: java.lang.IllegalArgumentException

+++ src/main/java/org/pengwt/myhome/myhome2/entity/BookMark.java
@@ -18,11 +18,11 @@ public class BookMark<T> {
     private long id;
     private String name;
     private String url;
-    private int userid;
+    private long userid;
 
-    public BookMark(String name, String url,int userid){
-    this.setName(name);
-    this.setUrl(url);
-    this.setUserid(userid);
+    public BookMark(String name, String url, long userid) {
+        this.setName(name);
+        this.setUrl(url);
+        this.setUserid(userid);
     }
 }
